Moved auxiliary types into their own dedicated files and group
